treewide: Fix default num of AXI-REALM managers and regenerate regs
This parameter in the makefrag affects the number of registers generated
inside the AXI-REALM IP. The default value, associated with the default
configuration of Cheshire, is 6, accounting for: `core0`, `debug
module`, `dma`, `serial link`, `vga` and `usb`. Change accordingly if
the configuration changes, e.g., the number of cores or if new managers
are added.